pyproject.toml
[tool.poetry]
name = "marketing-crew"
version = "0.1.0"
description = ""
authors = ["Kulwant Yadav kulwant209@gmail.com"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
crewai = "^0.11.0"
python-dotenv = "1.0.0"
opencv-python = "4.8.1.78"
google-search-results = "2.4.2"
newspaper3k = "^0.2.8"
lxml-html-clean = "^0.2.0"
flask = "^2.3.2"
setuptools = "^72.2.0"
unstructured = "^0.15.3"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
